move copy and data to parent class; put custom logic in pow
